string transformation file separated
