+++ bin/abell.js
@@ -4,7 +4,7 @@
  * It is very unlikely that you would have to change anything in this file
  * since it only deals with exporting right functions to users.
  *
- * You can check `src` folder for actual logic code
+ * You can check `src/cli.js` for actual logics.
  */
 
 // This is the file that gets called first on `abell [command]`
