+++ src/utils/helpers.js
@@ -5,16 +5,19 @@ const path = require('path');
  * @typedef {import('./typedefs').AbellConfigs}
  */
 
-const relativeJoinedPath = (pathString) =>
+/**
+ * Returns full path on given relative path
+ * @param {String} pathString - path object
+ * @return {String} relativelyJoinedPath
+ */
+const getAbsolutePath = (pathString) =>
   path.join(process.cwd(), ...pathString.split('/'));
 
-const getDirectories = (source) => {
-  return fs.readdirSync(source).filter((dirent) => {
-    return fs.lstatSync(path.join(source, dirent)).isDirectory();
-  });
-};
-
-const rmdirRecursiveSync = function (pathToRemove) {
+/**
+ * Removes the folder
+ * @param {String} pathToRemove path to the directory which you want to remove
+ */
+function rmdirRecursiveSync(pathToRemove) {
   if (fs.existsSync(pathToRemove)) {
     fs.readdirSync(pathToRemove).forEach((file, index) => {
       const curPath = path.join(pathToRemove, file);
@@ -28,9 +31,22 @@ const rmdirRecursiveSync = function (pathToRemove) {
     });
     fs.rmdirSync(pathToRemove);
   }
-};
+}
 
-const recursiveFindFiles = (base, ext, inputFiles, inputResult) => {
+/**
+ *
+ * @param {String} base directory to search in
+ * @param {String} ext Extension you want to search for (e.g. '.abell')
+ * @param {String[]} inputFiles Array of directories
+ * @param {String[]} inputResult Holds the old input result
+ * @return {String[]} Array of filepaths that end with given extension
+ */
+function recursiveFindFiles(
+  base,
+  ext,
+  inputFiles = undefined,
+  inputResult = undefined
+) {
   const files = inputFiles || fs.readdirSync(base);
   let result = inputResult || [];
 
@@ -51,9 +67,10 @@ const recursiveFindFiles = (base, ext, inputFiles, inputResult) => {
   }
 
   return result;
-};
+}
 
 /**
+ * Reads and returns content of abell.config.js
  * @return {AbellConfigs}
  */
 function getAbellConfigs() {
@@ -98,9 +115,9 @@ function getAbellConfigs() {
     abellConfig = defaultConfigs;
   }
 
-  const destinationPath = relativeJoinedPath(abellConfig.destinationPath);
-  const sourcePath = relativeJoinedPath(abellConfig.sourcePath);
-  const contentPath = relativeJoinedPath(abellConfig.contentPath);
+  const destinationPath = getAbsolutePath(abellConfig.destinationPath);
+  const sourcePath = getAbsolutePath(abellConfig.sourcePath);
+  const contentPath = getAbsolutePath(abellConfig.contentPath);
 
   return {
     ...abellConfig,
@@ -110,7 +127,11 @@ function getAbellConfigs() {
   };
 }
 
-const createPathIfAbsent = (pathToCreate) => {
+/**
+ * Recursively creates the path
+ * @param {String} pathToCreate path that you want to create
+ */
+function createPathIfAbsent(pathToCreate) {
   // prettier-ignore
   pathToCreate
     .split(path.sep)
@@ -121,7 +142,7 @@ const createPathIfAbsent = (pathToCreate) => {
       }
       return currentPath;
     }, '');
-};
+}
 
 /**
  *
@@ -200,84 +221,6 @@ const execRegexOnAll = (regex, template) => {
   return { matches: allMatches, input };
 };
 
-/**
- * Prefetchs links from given template and adds it to next template.
- * @param {Object} options
- * @param {String} options.from String of HTML/Abell template to fetch links from
- * @param {String} options.addTo String of HTML/ABELL template to add prefetch into
- *
- * @return {String}
- */
-function prefetchLinksAndAddToPage({ from, addTo }) {
-  const pageTemplate = addTo;
-
-  // eslint-disable-next-line
-  const regexToFetchPaths = /(?:<link +?rel=["']stylesheet['"] +?href=['"](.*?)['"])|(?:<script +?src=['"](.*?)['"])|(?:<link.+?href=["'](.*?)["'].+?as=["'](.*?)["'])/gs;
-  const { matches } = execRegexOnAll(regexToFetchPaths, from);
-  const headEndIndex = pageTemplate.indexOf('</head>');
-  if (headEndIndex < 0) return pageTemplate; // does not have </head>
-  // prettier-ignore
-  const newPageTemplate =
-    pageTemplate.slice(0, headEndIndex) +
-    `  <!-- Abell prefetch -->\n` +
-    matches
-      .map((link) => {
-        let stylesheet;
-        let script;
-        // stylesheet or script have a value if link is straighforward
-        // (e.g <link rel="stylesheet" href="style.css">)
-        ([stylesheet, script] = link.slice(1));
-        // In some cases, user may have a little trickier links
-        // (e.g <link rel="preload" href="next.js" as="script")
-        if (!stylesheet && !script) {
-          try {
-            if (link[4] === 'style' && link[3].includes('.css')) {
-              stylesheet = link[3];
-            } else if (link[4] === 'script' && link[3].includes('.js')) {
-              script = link[3];
-            }
-          } catch (err) {
-            console.log(">> Could not recognize preloads, skipping the option..."); // eslint-disable-line max-len
-          }
-        }
-        if (stylesheet) {
-          return `  <link rel="prefetch" href="${stylesheet.replace('../','./')}" as="style" />`; // eslint-disable-line max-len
-        } else if (script) {
-          return `  <link rel="prefetch" href="${script.replace('../', './')}" as="script" />`; // eslint-disable-line max-len
-        }
-      })
-      .join('\n') +
-    '\n\n' +
-    pageTemplate.slice(headEndIndex);
-
-  return newPageTemplate;
-}
-
-/**
- *
- * @param {String} htmlTemplate
- * @param {String} prefix
- * @return {String}
- */
-function addPrefixInHTMLPaths(htmlTemplate, prefix) {
-  const { matches, input } = execRegexOnAll(
-    / (?:href|src)=["'`](.*?)["'`]/g,
-    htmlTemplate
-  );
-
-  let output = '';
-  let lastIndex = 0;
-  for (const match of matches) {
-    if (match[1].startsWith('http') || match[1].startsWith('//')) continue;
-    const indexToAddOn = match.index + match[0].indexOf(match[1]);
-    output += input.slice(lastIndex, indexToAddOn) + prefix + '/';
-    lastIndex = indexToAddOn;
-  }
-  output += input.slice(lastIndex);
-
-  return output;
-}
-
 const boldRed = (message) =>
   `\u001b[1m\u001b[31m${message}\u001b[39m\u001b[22m`;
 const boldGreen = (message) =>
@@ -286,7 +229,6 @@ const grey = (message) => `\u001b[90m${message}\u001b[39m`;
 const yellow = (message) => `\u001b[1m\u001b[33m${message}\u001b[39m\u001b[22m`;
 
 module.exports = {
-  getDirectories,
   rmdirRecursiveSync,
   recursiveFindFiles,
   getAbellConfigs,
@@ -294,8 +236,6 @@ module.exports = {
   copyFolderSync,
   exitHandler,
   execRegexOnAll,
-  prefetchLinksAndAddToPage,
-  addPrefixInHTMLPaths,
   boldGreen,
   boldRed,
   grey,
